requirements.txt
discord.py
flask
pynacl